+++ src/mass_driver/commands.py
@@ -4,7 +4,12 @@ import os
 import sys
 from argparse import Namespace
 
-from mass_driver.discovery import discover_drivers, get_driver_entrypoint
+from mass_driver.discovery import (
+    discover_drivers,
+    discover_forges,
+    get_driver_entrypoint,
+    get_forge_entrypoint,
+)
 from mass_driver.main import main
 from mass_driver.migration import load_migration
 
@@ -29,6 +34,29 @@ def drivers_command(args: Namespace):
     print("Available drivers:")
     for driver in drivers:
         print(f"{driver.name}")
+    return
+
+
+def forges_command(args: Namespace):
+    """Process the CLI for 'Forges' subcommand"""
+    if args.info:
+        target_forge = args.info
+        try:
+            forge = get_forge_entrypoint(target_forge)
+            print(
+                f"Plugin name: {forge.name}; Import path: {forge.module}; Class: {forge.attr}"
+            )
+            print(forge.load().__doc__)
+            return
+        except ImportError as e:
+            print(str(e), file=sys.stderr)
+            print("Try `mass driver forges --list`", file=sys.stderr)
+            return
+    # if args.list:  # Implicit
+    forges = discover_forges()
+    print("Available forges:")
+    for forge in forges:
+        print(f"{forge.name}")
     return