+++ src/mass_driver/discovery.py
@@ -2,6 +2,7 @@
 
 import sys
 
+from mass_driver.forge import Forge
 from mass_driver.patchdriver import PatchDriver
 
 if sys.version_info < (3, 10):
@@ -35,3 +36,23 @@ def get_driver(driver_name: str) -> type[PatchDriver]:
     """Get the given driver Class, by entrypoint name"""
     driver = get_driver_entrypoint(driver_name)
     return driver.load()
+
+
+def discover_forges() -> EntryPoints:
+    """Discover all Forges via plugin system"""
+    return entry_points(group=FORGE_ENTRYPOINT)
+
+
+def get_forge_entrypoint(forge_name: str) -> EntryPoint:
+    """Fetch the given forge Entrypoint, by name"""
+    forges = discover_forges()
+    if forge_name not in forges.names:
+        raise ImportError(f"Forge '{forge_name}' not found in '{FORGE_ENTRYPOINT}'")
+    (forge,) = forges.select(name=forge_name)
+    return forge
+
+
+def get_forge(forge_name: str) -> type[Forge]:
+    """Get the given forge Class, by entrypoint name"""
+    forge = get_forge_entrypoint(forge_name)
+    return forge.load()